Fix exception encountered with ratelimiter
This issue appears to be due to the fact that the flask
dependency does not lock its own dependencies, causing
us to now pick up a more recent version of a library that
breaks our hard-coded version of Flask.
This is a workaround until we test and upgrade our Flask version
to a more recent release.
VIZOR-254