Temporary inefficient fix for UpdateInstitutionUsageStatsTask
VIZOR-334
While inefficent, this change to the query and
application-level table filtering allows us to workaround
modeling in the Cassandra database which is no longer
feasible for us to query for reads.

Future development should look at remodeling tables and / or
adding query tables optmized for this task. But the current fix
should allow us to run these pipelines again without failure
on the last task.