Undo previous attempt at data remodeling
While the remodel of the data table worked in single-node
local environments, it turned out to be problematic
in the three-node cluster deployed to production.
Additionally, we realized late that remodeling this table
would have implications that required us to modify the
tableau datasources and potentially workbooks.
So this change was backed out of the database.
This commit reverts the modeling change, but leaves its trace
in the git repository for posterity.
VIZOR-334