Integrate POC upload-ux into django application
VIZOR-363
VIZOR-364
VIZOR-369
VIZOR-377

This commit integrates a basic react POC into the django
ivweb application, with code directory organization and
build scripts.

The reasoning for committing the built React files is to
avoid the need to overhaul the existing deployment tooling.

The Vizor Manager application was originally written many years
ago to use bower, an old version of Node.js / npm, and
the Grunt task runner to build out its client-side assets.

This legacy tooling will not build a newer front end
application or component library. And a modern of Node.js
is not available on our iv-prod-admin VM for deployment
purposes.

Simply checking in the built react code will also give us a
record / copy of the file if some years down the line we
have a similar problem building older code.